// src/interfaces/error.interface.ts
export enum Error {
    IS_NOT_STRING = 1001,
    IS_NOT_NUMBER = 1002,
    IS_NOT_ARRAY = 1003,
    IS_NOT_OBJECT = 1004,
    IS_NOT_MIN_ARRAY = 2001,
    IS_NOT_MAX_ARRAY = 2001,
    IS_NOT_TITLE = 3001,
    IS_NOT_DESCRIPTION = 3002,
    IS_NOT_SELECT_CHANNEL = 4001,
    IS_NOT_INPUT_REASON = 4002,
    IS_NOT_ALLOW_TYPE = 4003,
    IS_NOT_SELECT_TYPE = 4004,
}